+++ app/src/main/java/woowacourse/omok/domain/board/Point.kt
@@ -1,0 +1,6 @@
+package woowacourse.omok.domain.board
+
+data class Point(
+    val x: Int,
+    val y: Int,
+)
