+++ app/src/main/java/woowacourse/omok/domain/board/StoneColor.kt
@@ -1,0 +1,13 @@
+package woowacourse.omok.domain.board
+
+enum class StoneColor {
+    WHITE,
+    BLACK,
+    NONE,
+    ;
+
+    fun reverseStoneColor(): StoneColor {
+        if (this == NONE) return NONE
+        return if (this == WHITE) BLACK else WHITE
+    }
+}
