+++ app/src/main/java/woowacourse/omok/domain/board/result/Finished.kt
@@ -1,0 +1,13 @@
+package woowacourse.omok.domain.board.result
+
+import woowacourse.omok.domain.board.Point
+
+sealed class Finished : PlaceStoneResult {
+    data class GameFinished(
+        val point: Point,
+    ) : Finished()
+
+    data class BoardFull(
+        val point: Point,
+    ) : Finished()
+}
