+++ app/src/main/java/woowacourse/omok/domain/board/result/OnGoing.kt
@@ -1,0 +1,15 @@
+package woowacourse.omok.domain.board.result
+
+import woowacourse.omok.domain.board.Point
+
+sealed class OnGoing : PlaceStoneResult {
+    data class StonePlaced(
+        val point: Point,
+    ) : OnGoing()
+
+    data object RuleViolation : OnGoing()
+
+    data object AlreadyPlaced : OnGoing()
+
+    data object InvalidMove : OnGoing()
+}
