+++ app/src/main/java/woowacourse/omok/domain/board/result/PlaceStoneResult.kt
@@ -1,0 +1,3 @@
+package woowacourse.omok.domain.board.result
+
+interface PlaceStoneResult
