+++ app/src/main/java/woowacourse/omok/domain/rule/count/OverlineRule.kt
@@ -1,0 +1,15 @@
+package woowacourse.omok.domain.rule.count
+
+import woowacourse.omok.domain.board.Board
+import woowacourse.omok.domain.board.Point
+
+class OverlineRule : OmokCountRule() {
+    override fun isCheckCondition(
+        board: Board,
+        previousPoint: Point,
+        dir: Pair<Int, Int>,
+    ): Boolean {
+        val total = checkDirection(board, previousPoint, dir)
+        return total > OMOK_COUNT
+    }
+}
