--- app/src/main/java/woowacourse/omok/model/board/Point.kt
@@ -1,6 +1,0 @@
-package woowacourse.omok.model.board
-
-data class Point(
-    val x: Int,
-    val y: Int,
-)
