--- app/src/main/java/woowacourse/omok/model/board/result/PlaceStoneResult.kt
@@ -1,3 +1,0 @@
-package woowacourse.omok.model.board.result
-
-interface PlaceStoneResult
