--- app/src/main/java/woowacourse/omok/model/rule/GameRule.kt
@@ -1,11 +1,0 @@
-package woowacourse.omok.model.rule
-
-import woowacourse.omok.model.board.Board
-import woowacourse.omok.model.board.Point
-
-interface GameRule {
-    fun validateMove(
-        board: Board,
-        validationPoint: Point,
-    ): Boolean
-}
