--- app/src/main/java/woowacourse/omok/model/rule/count/OverlineRule.kt
@@ -1,15 +1,0 @@
-package woowacourse.omok.model.rule.count
-
-import woowacourse.omok.model.board.Board
-import woowacourse.omok.model.board.Point
-
-class OverlineRule : OmokCountRule() {
-    override fun isCheckCondition(
-        board: Board,
-        previousPoint: Point,
-        dir: Pair<Int, Int>,
-    ): Boolean {
-        val total = checkDirection(board, previousPoint, dir)
-        return total > OMOK_COUNT
-    }
-}
